Finesse the behavior of Jsonnet auto-close symbols
vscode allows extension writers to customize "autoclose", e.g., when we
type the `{` character, we might wish for vscode to insert a
corresponding character `}` after the cursor.

This commit will change the auto-close behavior for the characters `{`,
`[`, and `'`.

For the case of `{` and `[`, this is straightforward: we close these
with `},` and `],`, respectively; the comma is included because Jsonnet
optionally allows commas after objects and lists in the vast majority of
circumstances, and it's more convenient to include it than to have the
user type it.

The `'` character is slightly more subtle. Currently if we're in a
comment, and we type the single quote character, vscode will add another
to close it. Normally this is useful, but in this case, it's vastly more
likely that we're using the single quote to write an English contraction
than anything else, and in this case, we do not want to close the quote.