Allow users to specify custom path for `jsonnet` binary
To allow users to generate a "preview" JSON for some Jsonnet file, this
extension shells out to the `jsonnet` command line utility. Currently we
depend on users having this on the path (e.g., by `brew install`'ing
it).

This commit adds a configurable setting, `jsonnet.executablePath`, that
users can customize in their `settings.json`, which allows them to point
at a custom `jsonnet` binary.