Fix broken image link on vscode extension marketplace
Images with relative paths are broken when displayed on the vscode
marketplace website, because it doesn't host the images. Currently we
are using a relative path for our main demo image, which causes it to be
broken on the marketplace page for this extension. This commit will fix
this by adding an absolute URL pointing at the image.