Add simple partial AST implementation and visitor
A Jsonnet language server must be able to deserialize an AST emitted by
a code provider. Since our code provider emits JSON-serialized ASTs,
this commit will introduce a small number of JSON-serializable node
classes that will be used internally to represent the AST.

Additionally, we introduce an AST visitor, which will handle a lot of
the AST traversal logic.