Implement location-aware object fields
When a user hovers the mouse cursor over a symbol, we would like to
resolve the symbol and report documentation for it. This implies being
able to accurately understand where the cursor currently exists in the
source.

Since objects can become quite large, this commit will add location data
to the object field AST node.