Use TypeScript's strict nullable checking
The TypeScript compiler has a useful feature that allows you to require
that types are explicitly marked as nullable if you want to assign a
null value of some sort to them.

This commit will turn this flag on for our codebase. This decision
carries with it a number of benefits:

* For types that are deserialized, this clearly communicates what fields
  we expect to sometimes be null (i.e., which fields are implemented as
  pointers in the Go version of the AST)
* The compiler will warn you if you're using a nullable field without
  checking that it could be null.
* When annotated correctly, this prevents null pointer exceptions from
  being thrown anywhere in your code (though, of course, libraries are
  another story).

Part of this decision necessarily involves structuring the code so that
we are checking for `null` when accessing fields that are nullable. In
addition to this, one of the choices we have made here is to not make
any of our _types_ nullable, instead opting to make _variables_
nullable. The reasoning for this is (1) making a type nullable will have
a cascading effect across the codebase, causing every callsite that uses
a variable of that type to do a null check, and (2) there don't seem to
be any good reasons to justify making a type always nullable.