Move traversal logic to switch statement in `Visit`
Currently every visitor that implements the `Visitor` interface needs to
define how children are visited (i.e., what order, which children, and
so on).

A better way is to define the traversal in `Visitor.VisitorBase`, and
then simply have overriding methods define what to do when a node is
encountered.