Add `parent` field to `ast.Node` interface
This commit will introduce a field, `parent` to the `ast.Node`
interface. This field is particularly useful when doing static analysis,
as it allows the user to move freely through the tree.

This change is accompanied with a change to how the visitor works.
Instead of visitor overloads returning an object of type `T`, they all
return `void`. With this change, results should be obtained with
properties on the visitor. The advantage to this approach is that it
simplifies the use of inheritance, which in JavaScript is somewhat
clumsy and unintuitive.