Transition away from dynamically-scoped `this`
In JavaScript (and, hence, TypeScript), `this` is dynamically scoped.
This is well understood and documented, so rather than explain it here,
we will simply say that it leads to unintuitive behavior, and that the
goal of this behavior is to remove all opportunities for it to affect
our code.

To accomplish this, we observe that TypeScript's so-called "fat arrow"
functions implement lexically-scoped `this`, and hence this commit will
replace all traditional functions with their fat arrow equivalents.