Cause `analyzer.parseJsonnet` to populate parent and env
Some information that exists on the AST (e.g., the environment, parent
information for each node) is expected to be populated by visitors.
Currently, though, not all parsing code uses a visitor to populate this
information on the AST, with the consequence that code that depends on
such information will not work correctly.

This commit will address this by introducing a visitor class
specifically for parsing code to generate this information before
returning an AST.