[WIP] Thread lexer through completion provider
Currently the completion provider always returns the same 3 (bogus)
suggestions. This commit begins the process of generating useful
completion suggestions, by lexing input leading in the vicinity of the
cursor when the completion event is fired, and threading that through to
the site where we will eventually replace the bogus completion
suggestions.

The next step is to parse the lexed input, and attempt to resolve the
symbols being typed against the current environment.