[WIP] Proof-of-concept quality autocomplete
This commit will add PoC-quality autocomplete to the vscode extension.
This work is meant to bring us to a place where we can write unit tests
against something that works, and incrementally bring the extension to
production quality. Because this code is not (even putting it
charitably) good, the next step of the dev process is to take this code
and turn it into something maintainable.