[WIP] Begin to transition language server to be interface-basedh
Currently, `ast`, the folder with code responsible for static analysis
features, is a monolith. If we are to extend this codebase to support
code completion over (e.g.) HTTP, as we would need in the short term to
enable demos in the browser, we would need to start abstracting out the
functionality into interfaces that can be implemented in a variety of
contexts.

This commit will begin this process by starting to transition the
evented code over to `AnalysisEventListener`, which abstracts the idea
of analysis features like `onHover` events.