[WIP] Move `onComplete` into `AnalysisEventListener` interface
This commit will continue the work of breaking the language service down
into interfaces that can be mocked, extended, and re-used in a variety
of use cases (e.g., having a TypeScript frontend talk to a compiler over
HTTP).

This commit will specifically move the meaty chunk of code required for
the `onComplete` event into the interface we use for evented analysis,
called `AnalysisEventListener`.