Move workspace interfaces to their own file
Currently `ast/analyzer.ts` contains a hodge-podge of interfaces
representing different aspects of the problem of creating a Jsonnet
language service.

In this commit we take our first step to correcting this, separating out
the interfaces that are responsible for managing the workspace (e.g.,
`DocumentManager`, which abstracts over vscode's `TextDocuments`, which
is the system of record for all documents in the workspace, and which
emits events when any document is changed), and placing them in their
own file.