[WIP] Introduce compiler services interfaces and implementations
Currently the `Analyzer` class handles both the logic of registering
hooks to process events that are fired off in response to changes to the
workspace (e.g., in responce to a document save event, a `Hover` event),
as well as the tasks of maintaining a cache of syntactically analyzed
text.

One of our uses cases is for this code to be running in a browser,
talking to the compiler service over HTTP. To make room for this, we
must break the compiler service into an interface that the `Analyzer`
can consume.

This commit will introduce such interfaces, and begin the process of
transitioning the `Analyzer` class to using an implementation of them
that simply shells out to a local compiler toolchain.