Implement Jsonnet parser in pure TypeScript
This commit will introduce a (very hacky) Jsonnet parser, written and
tested in pure TypeScript. This primarily involves four major work
items:

1. Implementing the complete token and AST API. This task was about 10%
   done already. The bulk of the remaining work lies in defining
   interfaces and classes that capture the structure of these APIs.
2. Implementing the lexer and the parser from scratch.
3. Porting the tests.
4. Transitioning the whole codebase to use the TS parsing stack, instead
   of shelling out to a language server.

At this point, this code brings us to feature parity with the previous
version of the code, which depended on an external binary shipping
serialized ASTs back to the TypeScript extension.

Some limitations of the code include:

* Does not handle some features, like default params.
* Code in the analyzer is still only partially implemented; this commit
  only implements a Jsonnet parser, it does not change the architecture
  of the existing analysis code.