Transition away from the pattern of aliasing `this`
Currently the lexer and parser use a very strange pattern, which is to
define an alias for this, as in:

    const l = this; // l is a lexer

This pattern is an artifact of the port from Go. We meant to clean it
up, but never got around to it.

This commit will transition the parser and lexer away from this pattern,
opting instead for simply using `this`.