Implement analyzer using the TypeScript parser
As the user edits a Jsonnet document, it is left in a state of perpetual
brokenness. The analyzer attempts to maintain enough context that it can
offer useful suggestions even when the document (temporarily) doesn't
parse.

Currently the analyzer implements this "partial parse" functionality by
lexing the edited document, and attempting to make sense of the tokens
in the immediate vicinity of the cursor.

This is obviously problematic because the context of the AST is lost
when we only look at tokens. For example, if cursor points at the name
in a variable definition, obviously we do not want to offer an
autocompletion.

This commit begins the process of fixing this by transitioning the
analyzer away from dealing with lexer tokens, opting instead to
implement the analysis logic in terms of the AST output by the Jsonnet
parser that we have written in TypeScript.