Implement default parameter values in Jsonnet parser
Jsonnet supports a notion of function parameters having default values.
For example:

    local foo(bar=4) = bar; foo(1)

Currently this is not implemented in our Jsonnet parser. This commit
will add this functionality to the parser.

Since parameters in function definitions are currently represented using
the node type `IdentifierName` (which is currently implemented as an
alias for `string`), the work here largely falls into defining a new
node type, `FunctionParam`, which contains a field for the (optional)
default value of the parameter, and phasing out the use of
`IdentifierName` as a means for representing parameters in function
definitions.