Add type guards for every interface that extends `Node`
This commit allows us to use pattern matching-like facilities on all
types that extend `Node`. This is implemented using TypeScript's
user-defined type guards, which is meant to allow users to easily deal
with union types (e.g., `foo | bar`). Essentially, the idea is to be
able to use an `if` to automatically infer the runtime type of an
object, like:

  const someObj: foo | bar = fooOrBar();
  if (isFooInstance(someObj)) {
    // TypeScript knows that `someObj` is of type `foo`, and allows us
    // to call `foo`'s `fooFunc` method.
    someObj.fooFunc();
  }
  // TypeScript knows that `someObj` is either a `foo` or a `bar`, and
  // since it's not `foo`, it must be `bar`. So, it allows us to call
  // the `bar`'s `barFunc` on `someObj`.
  someObj.barFunc();

In order to get this sort of inferred "type switching" behavior, we
define a type guard, like:

  export const isFooInstance = (a: any): a is foo => {
    return a instanceof foo;
  }

By defining such a type guard for every `Node` type, we are able to get
similar behavior for all `Node`s.