[Parser] Add object's `local` fields to scope before we visit fields
Suppose we have a Jsonnet template:

  {
    bar: {baz: foo},
    local foo = 3,
  }

Because `local` field definitions are order-independent, `foo` is
actually in scope when it is used in the definition of the field `bar`.

Currently our compiler will not correctly populate these definitions
into the environment before visiting the definitions of object fields,
which means that our static analyzier cannot resolve the reference to
`foo` when a user (e.g.) hovers over it.

This commit will fix that by collecting `local` fields and adding them
to the environment before visiting object fields.