[Analyzer] Allow resolving symbols from files that `import`
Currently the analyzer will choke if we try to resolve a symbol from a
file that has an `import` at the top. For example, if we have some index
expression `foo.bar`, and `foo` is an import of a file that looks like
this:

  local thing = import "some file";
  {
    bar: 3,
  }

Then currently the analyzer will fail when resolving `bar`. This is
because the import is resolved to a `Local` node, but the analyzer
expects an `ObjectNode`.

This commit will resolve this issue by stripping `Local`s off an
imported file when we attempt to resolve an index expression.