[Parser] Reorganize object hierarchy for `Node`
In this commit we will transition the `Node` hierarchy mostly to a class
hierarchy, primarily to pave the way to re-implement the `resolve*`
functions on the `Node` object, so that you can do something like
`someVarNode.resolve()`, which would try to resolve the variable and
return a value for it.

Additionally, we move pretty-printing to the `Node` classes, as this is
a core use case for us (i.e., we need to be able to render a line in a
tooltip when a user hovers over it).

A more technical treatment of this follows.

The parser currently implements the AST using a set of hierarchical
data-only interfaces (i.e., interfaces with no methods). For example,
the interface `ast.Binary` extends the base interface `ast.Node`, and
each contains only properties meant for data, rather than properties
that point at functions.

This has two big advantages.

First, the fact that the data objects are extremely lightweight (e.g.,
they have no methods, no private data fields, and so on) make it very
easy to reason about code. There are no prototypal inheritance chains,
no dynamic dispatch, no complex private data or constructor magic.
Essentially anything with the right fields is an instance of an
interface.

Second, Using data-only interfaces means that essentially any object
that happens to have the correct object fields implements the interface
incidentally. This is very convenient if (e.g.) you expect to serialize
to/from JSON, since we don't have to deal with issues like populating
methods on that interface.

But our use of them has also gotten complex enough that the data objects
are really not simple enough to map to such a simple data model anymore:

1. Things like pretty-printing (which is essential for our use case) and
   resolving symbols must be implemented in functions that are not
   directly attached to the interfaces, which is somewhat awkward.
2. We re-implement many things that are given to you "for free" by OOP.
   For example, the `make*` functions that act essentially as
   constructors for these interfaces, or the fact that we often
   instantiate objects that happen to implement an interface, which
   results in a bunch of duplicated code for things like setting `env`
   and `parent` to `null` initially.
3. Complicates user-defined type guards. TypeScript has a pattern
   matching-like feature called user-defined type guards; to use this
   feature, the user must write some code that tells the TypeScript
   compiler when an object is considered to be of a specific type. When
   the type is an interface, you must essentially verify that all the
   necessary fields are there, which is ungainly next to the alternative
   when we use classes, which is to use `instaceof`.

And so on.

Of course, none of these are an absolute deal-breaker, but ultimately
they add up to a situation where the utility of keeping an
interface-only data hierarchy is slightly more trouble than it's worth.