[Analyzer] Correctly resolve function parameters
Currently, if the user hovers over an identifier that should resolve to
a function parameter, we emit the "fallback" text, which is currently
the line of the cursor's current position. A better idea is to resolve
resolve and emit the `FunctionParam` that the identifier refers to.

This commit will implement such behavior.