[Analyzer] Add parameters of `local` to scope before visiting body
Currently if we attempt to resolve a variable that refers to a function
paramter of a `local` function, the analyzer will get confused and
throw. This is because we're not adding parameters of `local` functions
to scope before visiting the body, which causes us to be unable to
resolve the variable.

This commit will correct this error.