Remove (undocumented) vscode configuration variable, `jsonnet.server`
Previously this vscode extension depended on a Jsonnet language server,
which we would shell out to, in order to get static analysis information
back.

The variable that users would have used to configure this behavior was
never documented, and since implementing a Jsonnet parser in TypeScript,
we have no further need for it.

Thus, this commit removes the option altogether, with no deleterious
effects in functionality.