[Parser] Introduce "rest" member into static error
When a document doesn't parse, autocompletion is harder. Less context
means we have to try to recover more information "the hard way". To help
autocompltion along, it is frequently useful to emit a partial parse
result along with the error.

Borrowing the term from Lisp, this commit introduce a "rest" field to
the static error class, which contains the "rest" of the failed parse.
This will often contain things like a trailing index expression (say),
or a variable node. For example, when the user presses the `.` key, it
is quite likely the document will no longer parse, and when the parser
fails, it would emit the left-hand side of the index expression that
precedes the `.`.