Move symbol resolution code from analyzer to `Node` classes
Currently symbol resolution is implemented as an ad hoc series of static
functions that analyze the AST. This leaves us in an uncomfortable
position from a maintainability perspective: it is not clear which nodes
can be resolved, or what the semantics are of individual symbol
resolution functions are.

This commit will introduce an interface, `Resolvable`, which is meant to
be implemented by a subset of the `Node` classes (for example, `Var`,
`Identifier`, and so on). The semantics of `Resolvable#resolve` are to
produce either `null` (in the case we can't resolve the `Node`), or the
`Node` that the symbol would refer to, if it were to be evaluated.
Because we define a type guard for `Resolvable`, resolving a node is
often as easy as:

  if (isResolvable(someNode)) {
    someNode.resolve();
  }

This is markedly easier to reason about than the ad hoc `switch`
statments we employ throughout the code to customize resolution
behavior. This way is clearer because the user knows that the semantics
of resolution exist in the `Node` implementation.

This commit will also introduce another utility interface,
`FieldsResolvable`, which is somewhat like `Resolvable`, except it
attempts to determine what statically-knowable fields a `Node` has. This
is useful for (e.g.) suggesting fields for autocompletion.