Use `CompletionItemKind.Variable` for `local`s
In an `onCompletion` event, `local` variables are emitted with
completion item kind `Field`. This is the same kind as object fields,
which makes autocomplete menus a bit more dense to read, because they
all subsequently use the same icon.

This commit will transition `local` variables to use the `Variable` kind
instead, which helps to differentiate.