Implement document caching in `VsCompilerService`
Currently the `VsCompilerService` will always re-parse a document for
analysis when we call the `cache` method. This commit will introduce
caching semantics to that function. Specifically, we cache on the
document URI and an optional version. If the user passes `undefined` for
the version, (e.g., if it comes from a source that does not have a
version, like when we are pulling the file directly from the FS) we will
always parse from scratch.