[Analyzer] Return node for positions beyond the end of text
Currently if the user types a `.` character at the end of the document,
we'll fail to autocomplete. This is because the analyzer attempts to
find the node in whose range the cursor lies, but because the document
no longer parses (a trailing `.` is syntactically invalid), we must
consult the last good parse. But because that document was one character
shorter than the current, non-parsing document, we fail to find the node
whose range contains the cursor.

This commit will fix this by simply allocating cursors beyond the text
range to the last node in the tree.