Refactor the AST search and resolution routines
The AST searching and symbol resolution facilities were largely written
before we had a good understanding of the Jsonnet language and its
amenability to static analysis. As a result, many of our core
abstractions are fundamentally at odds with the goals of our static
analysis, and greatly complicate (e.g.) the `onComplete` logic.

This commit will begin the process of:

  1. breaking the result of a search in the AST into a more fine-grained
     success and failure cases (e.g., "we found a node, but it wasn't an
     identifier", "we failed to find a node in the AST that wraps the
     cursor location because the cursor lies outside the range of the
     line as recorded by the last good parse of the document", and so
     on).
  2. breaking the result of an attempt to resolve a symbol into more
     fine-grained success and failure cases (e.g., "we tried to resolve
     the node, but we couldn't because it is a free parameter").

Explicitly outside the scope of these changes is:

  1. cleaning up the logic of `onComplete`, which will be a challenge
     that deserves its own patch set.

Signed-off-by: Alex Clemmer <clemmer.alexander@gmail.com>