Allow node type guards to take `any` type as argument
Previous implementations of the `ast.Node` inheritance hierarchy used
`interface` rather than `class` almost exclusively. In this context, the
use of `instanceof` to check whether the parameter implements a type did
not make sense, because interfaces do not have constructors, so it is
impossible to create an instance of an interface.

As a result of this, the user-defined type guards for the `ast.Node`
hierarchy took specific `ast.Node` types as arguments, since we could
leverage the type checker to make sure that what was being passed in was
actually an `ast.Node` type.

Somewhere along the way, we changed the `ast.Node` hierarchy to use
classes, but we didn't understand enough about the TypeScript/JavaScript
model to know that it was ok to change the parameters these type guards
take to `any`, since an instance of a class can be checked with
`instanceof`.

This commit does exactly that, since we will benefit in the next few
commits from being able to use user-defined type guards on types that
arent in the `ast.Node` inheritance hierarchy.

Signed-off-by: Alex Clemmer <clemmer.alexander@gmail.com>