Remove spurious exceptions, and standardize message format
This commit will (1) standardize the exception message format used
throughout the codebase, and (2) remove the use of spurious exceptions.

The spurious exceptions almost entirely lie in null checks during the
use of `immutable.js`'s iterators. These iterators type the elements
we're iterating over as nullable, and since they should never be null,
we currently throw if we find that they are null.

This is not strictly necessary, though, since the elements we are
iterating over will never be null. We can hence avoid these checks by
specifying the type in the iterator manually to be non-nullable.

Signed-off-by: Alex Clemmer <clemmer.alexander@gmail.com>