Add `LocalBind` to the `ast.Node` inheritance hierarchy
Currently, binds in a `local` are represented with a helper class,
`LocalBind`. This class does not implement `ast.Node`, which leads to
several places in the codebase where `LocalBind` is a special case.

One particular example of this is the place in the analyzer's
`onComplete` logic, when we have to retrieve the environment of the
local bind to "glue together" a partial parse and a found node to obtain
sensible autocomplete suggestions in a document that doesn't parse --
since `LocalBind` does not inherit from `ast.Node`, we don't have an
`env` member to look up variables in, so we simply use the environment
of the `LocalNode` instead.

This commit will simplify a lot of this logic by simply making
`LocalBind` a part of the `ast.Node` hierarchy.

Signed-off-by: Alex Clemmer <clemmer.alexander@gmail.com>