Add support for customizing Jsonnet library search path
The Jsonnet CLI allows users to customize which paths to search for
Jsonnet libraries, using the `-J` "jpath" flag.

This commit will add a similar feature to the vscode extension, exposing
a configuration property, `jsonnet.libPaths`, which is an array of
string that can be set for either a workspace or a user, in vscode
`settings.json` files.

Signed-off-by: Alex Clemmer <clemmer.alexander@gmail.com>