[1/2] Emit warning if extension can't find an import path
This change is split into two commits, as this changeset breaks
`onHover` resolution, while the next changeset fixes it.

Currently if an import points at an invalid path, the user would not
know unless they tried to compile or use autocomplete (which would
mysteriously fail to emit sensible suggestions).

This commit introduces:

1. A diagnostic visitor implementation, which will traverse a Jsonnet
   AST looking for diagnostic issues to report to the user, and
2. A feature in the diagnostic visitor that checks the paths we
   `import`, and emits a warning telling them we couldn't find the file
   they're referencing, and they may need to set `jsonnet.libPaths` if
   they haven't.

An ancillary benefit of this feature is that it should guide new users
to a scenario where they are more likely to get autocomplete
functionality.

Signed-off-by: Alex Clemmer <clemmer.alexander@gmail.com>