Add `LibPathResolver` abstraction
Jsonnet formally supports a notion of "jpaths", which are paths the
Jsonnet compiler driver will look for Jsonnet libraries. Typically these
are specified with the `-J` flag on the Jsonnet CLI.

As we move towards supporting this concept formally, it will be useful
to abstract away the problem of searching a filesystem for libraries, in
a way that is mockable and testable.

To this end, this commit will introduce:

1. `LibPathResolver`, a abstract class that abstracts over the idea of
   searching a set of library paths, and
2. `VsPathResolver`, an implementation of `LibPathResolver` that
   implements the platform-specific parts for the context of running in
   vscode.

The strength of this approach is that `LibPathResolver` implements the
path searching logic in a way that only requires users to implement
`LibPathResolver#pathExists` in order to integrate with new filesystems
or mock filesystems.

Signed-off-by: Alex Clemmer <clemmer.alexander@gmail.com>