[2/2] Introduce resolution context, fix symbol resolution
Symbol resolution across imports is currently broken when resolving
through `import` statements.

The essence of the problem is that the analyzer will try to obtain a
document from a path, but the path is sometimes resolved by calling
`path.resolve` on filename found in the AST node.

This, of course, breaks in a few scenarios, particularly when we are
looking for a Jsonnet library on the `jpath`, which is not relative to
the current file.

This commit introduces a "resolution context" that tracks the current
file URI as we continue the process of symbol resolution. For example,
when we search `libPaths` for a file, and we find it, we will want to
propagate the path we found it at, through the resolution code, so that
we can correctly cache it, etc.

Signed-off-by: Alex Clemmer <clemmer.alexander@gmail.com>