Don't use `backup` in `peek`
`peek` is a traditional parsing routine that allows you to inspect, but
not consume, the next token in a token stream.

Currently `peek` is implemented by popping the token, calling `backup`,
and returning the token. This causes problems if you want to call
`backup` later, because `backup` can only be called once, to make sure
parsing is at least reasonably performant.

This commit will re-implement `peek` so that it simply looks at the last
token on the stack, rather than calling `next` and `backup`.

Signed-off-by: Alex Clemmer <clemmer.alexander@gmail.com>