Treat block comments separated by blank lines as separate blocks
In the current analyzer, two C-style or hash-style comment blocks
separated by a newline are treated as one block. For example:

  {
    // This comment should not appear `onHover`.

    // This comment SHOULD appear `onHover`.
    foo: "bar",
  }

In the current analyzer, the comment resolved for `foo` would include
both of these lines.

This commit will transition to a parsing regiment that propertly return
only the second comment.

Signed-off-by: Alex Clemmer <clemmer.alexander@gmail.com>