Register Jsonnet runtime errors with vscode
Currently the vscode extension will render errors and warnings for
things that can be determined statically, such as syntax errors and
failure to find imports.

This commit will expand this capability to include Jsonnet runtime
issues. The workflow is as follows:

  1. The user types some code.
  2. The user clicks "save".
  3. The vscode extension compiles the Jsonnet code to JSON.
  4. If the compilation fails, the problematic lines are highlighted in
     the editor.

By default, this will always happen. We do not yet support turning off
this feature.

Signed-off-by: Alex Clemmer <clemmer.alexander@gmail.com>